+++ src/CTLD_i18n_es.lua
@@ -10,7 +10,7 @@ if not ctld then ctld = {} end
 if not ctld.i18n then ctld.i18n = {} end
 
 ctld.i18n["es"] = {}
-ctld.i18n["es"].translation_version = "1.11"
+ctld.i18n["es"].translation_version = "1.12"
 
 --- groups names
 ctld.i18n["es"]["Standard Group"] = "Grupo estándar"
@@ -556,3 +556,6 @@ ctld.i18n["es"]["Ground Vehicles"] = ""
 ctld.i18n["es"]["Helicopters"] = ""
 ctld.i18n["es"]["Infantry"] = ""
 ctld.i18n["es"]["Ships"] = ""
+
+--- Keys added by generate_i18n_dicts.ps1 on 2026-07-30
+ctld.i18n["es"]["%1 setting(s) absent from the mission config — CTLD default used: %2"] = ""