+++ src/CTLD_i18n_ko.lua
@@ -10,7 +10,7 @@ if not ctld then ctld = {} end
 if not ctld.i18n then ctld.i18n = {} end
 
 ctld.i18n["ko"] = {}
-ctld.i18n["ko"].translation_version = "1.11"
+ctld.i18n["ko"].translation_version = "1.12"
 
 --- groups names
 ctld.i18n["ko"]["Standard Group"] = "표준 그룹"
@@ -425,3 +425,6 @@ ctld.i18n["ko"]["Ground Vehicles"] = ""
 ctld.i18n["ko"]["Helicopters"] = ""
 ctld.i18n["ko"]["Infantry"] = ""
 ctld.i18n["ko"]["Ships"] = ""
+
+--- Keys added by generate_i18n_dicts.ps1 on 2026-07-30
+ctld.i18n["ko"]["%1 setting(s) absent from the mission config — CTLD default used: %2"] = ""